+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The corresponding alert states are:
 ### Features
 
 - [#740](https://github.com/influxdata/kapacitor/pull/740): Support reset expressions to prevent an alert from being lowered in severity. Thanks @minhdanh!
+- [#670](https://github.com/influxdata/kapacitor/issues/670): Add ability to supress OK recovery alert events.
 
 ### Bugfixes